requirements.txt
faker>=20.0.0
python-dateutil>=2.8.0 